Refactor: refatorando para implementar auth
