+++ app/Http/Requests/Api/V1/Category/CategoryRequest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Http\Requests\Api\V1\Category;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class CategoryRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return auth()->check();
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'name' => 'required|string|unique:categories,name|min:3',
+        ];
+    }
+}
