+++ tests/Helpers/AuthHelper.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\Helpers;
+
+use App\Models\User;
+use Laravel\Sanctum\Sanctum;
+
+class AuthHelper
+{
+    public static function authenticate()
+    {
+        $user = User::factory()->create();
+        Sanctum::actingAs($user);
+
+        return $user;
+    }
+}
